+++ Source/Scene/Scene.js
@@ -137,6 +137,7 @@ define([
         this._passState = new PassState(context);
         this._canvas = canvas;
         this._context = context;
+        this._globe = undefined;
         this._primitives = new CompositePrimitive();
         this._pickFramebuffer = undefined;
         this._camera = new Camera(this);
@@ -452,6 +453,22 @@ define([
         },
 
         /**
+         * Gets or sets the depth-test ellipsoid.
+         * @memberof Scene.prototype
+         * @type {Globe}
+         */
+        globe : {
+            get: function() {
+                return this._globe;
+            },
+
+            set: function(globe) {
+                this._globe = this._globe && this._globe.destroy();
+                this._globe = globe;
+            }
+        },
+
+        /**
          * Gets the collection of primitives.
          * @memberof Scene.prototype
          * @type {CompositePrimitive}
@@ -506,6 +523,31 @@ define([
             get : function() {
                 return this._animations;
             }
+        },
+
+        /**
+         * Gets the collection of image layers that will be rendered on the globe.
+         * @memberof Scene.prototype
+         * @type {ImageryLayerCollection}
+         */
+        imageryLayers: {
+            get : function() {
+                return this.globe.imageryLayers;
+            }
+        },
+
+        /**
+         * The terrain provider providing surface geometry for the globe.
+         * @memberof Scene.prototype
+         * @type {TerrainProvider}
+         */
+        terrainProvider: {
+            get : function() {
+                return this.globe.terrainProvider;
+            },
+            set : function(terrainProvider) {
+                this.globe.terrainProvider = terrainProvider;
+            }
         }
     });
 
@@ -513,11 +555,11 @@ define([
     var scratchOccluder;
 
     function getOccluder(scene) {
-        // TODO: The occluder is the top-level central body. When we add
+        // TODO: The occluder is the top-level globe. When we add
         //       support for multiple central bodies, this should be the closest one.
-        var cb = scene._primitives.centralBody;
-        if (scene.mode === SceneMode.SCENE3D && defined(cb)) {
-            var ellipsoid = cb.ellipsoid;
+        var globe = scene.globe;
+        if (scene.mode === SceneMode.SCENE3D && defined(globe)) {
+            var ellipsoid = globe.ellipsoid;
             scratchOccluderBoundingSphere.radius = ellipsoid.minimumRadius;
             scratchOccluder = Occluder.fromBoundingSphere(scratchOccluderBoundingSphere, scene._camera.positionWC, scratchOccluder);
             return scratchOccluder;
@@ -1039,6 +1081,10 @@ define([
         var frameState = scene._frameState;
         var commandList = scene._commandList;
 
+        if (scene._globe) {
+            scene._globe.update(context, frameState, commandList);
+        }
+
         scene._primitives.update(context, frameState, commandList);
 
         if (defined(scene.moon)) {
@@ -1370,10 +1416,10 @@ define([
      * @memberof Scene
      */
     Scene.prototype.morphTo2D = function(duration) {
-        var centralBody = this.primitives.centralBody;
-        if (defined(centralBody)) {
+        var globe = this.globe;
+        if (defined(globe)) {
             duration = defaultValue(duration, 2000);
-            this._transitioner.morphTo2D(duration, centralBody.ellipsoid);
+            this._transitioner.morphTo2D(duration, globe.ellipsoid);
         }
     };
 
@@ -1383,10 +1429,10 @@ define([
      * @memberof Scene
      */
     Scene.prototype.morphToColumbusView = function(duration) {
-        var centralBody = this.primitives.centralBody;
-        if (defined(centralBody)) {
+        var globe = this.globe;
+        if (defined(globe)) {
             duration = defaultValue(duration, 2000);
-            this._transitioner.morphToColumbusView(duration, centralBody.ellipsoid);
+            this._transitioner.morphToColumbusView(duration, globe.ellipsoid);
         }
     };
 
@@ -1396,10 +1442,10 @@ define([
      * @memberof Scene
      */
     Scene.prototype.morphTo3D = function(duration) {
-        var centralBody = this.primitives.centralBody;
-        if (defined(centralBody)) {
+        var globe = this.globe;
+        if (defined(globe)) {
             duration = defaultValue(duration, 2000);
-            this._transitioner.morphTo3D(duration, centralBody.ellipsoid);
+            this._transitioner.morphTo3D(duration, globe.ellipsoid);
         }
     };
 
@@ -1420,6 +1466,7 @@ define([
         this._screenSpaceCameraController = this._screenSpaceCameraController && this._screenSpaceCameraController.destroy();
         this._pickFramebuffer = this._pickFramebuffer && this._pickFramebuffer.destroy();
         this._primitives = this._primitives && this._primitives.destroy();
+        this._globe = this._globe && this._globe.destroy();
         this.skyBox = this.skyBox && this.skyBox.destroy();
         this.skyAtmosphere = this.skyAtmosphere && this.skyAtmosphere.destroy();
         this._debugSphere = this._debugSphere && this._debugSphere.destroy();