+++ Specs/Scene/SceneSpec.js
@@ -5,8 +5,10 @@ defineSuite([
          'Core/Color',
          'Core/Cartesian3',
          'Core/BoundingSphere',
+         'Core/Ellipsoid',
          'Core/Event',
          'Core/Rectangle',
+         'Renderer/ClearCommand',
          'Renderer/DrawCommand',
          'Renderer/Context',
          'Renderer/Pass',
@@ -16,12 +18,15 @@ defineSuite([
          'Renderer/UniformState',
          'Scene/AnimationCollection',
          'Scene/Camera',
+         'Scene/Globe',
          'Scene/CompositePrimitive',
          'Scene/RectanglePrimitive',
          'Scene/FrameState',
          'Scene/OIT',
          'Scene/ScreenSpaceCameraController',
          'Specs/createScene',
+         'Specs/equals',
+         'Specs/render',
          'Specs/destroyScene'
      ], 'Scene/Scene', function(
          defined,
@@ -29,8 +34,10 @@ defineSuite([
          Color,
          Cartesian3,
          BoundingSphere,
+         Ellipsoid,
          Event,
          Rectangle,
+         ClearCommand,
          DrawCommand,
          Context,
          Pass,
@@ -40,12 +47,15 @@ defineSuite([
          UniformState,
          AnimationCollection,
          Camera,
+         Globe,
          CompositePrimitive,
          RectanglePrimitive,
          FrameState,
          OIT,
          ScreenSpaceCameraController,
          createScene,
+         equals,
+         render,
          destroyScene) {
     "use strict";
     /*global jasmine,describe,xdescribe,it,xit,expect,beforeEach,afterEach,beforeAll,afterAll,spyOn,runs,waits,waitsFor,WebGLRenderingContext*/
@@ -400,6 +410,47 @@ defineSuite([
         destroyScene(s);
     });
 
+    it('setting a central body', function() {
+        var scene = createScene();
+        var ellipsoid = Ellipsoid.UNIT_SPHERE;
+        var globe = new Globe(ellipsoid);
+        scene.globe = globe;
+
+        expect(scene.globe).toBe(globe);
+
+        destroyScene(scene);
+    });
+
+    it('destroys primitive on set globe', function() {
+        var scene = createScene();
+        var globe = new Globe(Ellipsoid.UNIT_SPHERE);
+
+        scene.globe = globe;
+        expect(globe.isDestroyed()).toEqual(false);
+
+        scene.globe = null;
+        expect(globe.isDestroyed()).toEqual(true);
+
+        destroyScene(scene);
+    });
+
+    it('renders a central body', function() {
+        var s = createScene();
+
+        s.globe = new Globe(Ellipsoid.UNIT_SPHERE);
+        s.camera.position = new Cartesian3(1.02, 0.0, 0.0);
+        s.camera.up = Cartesian3.clone(Cartesian3.UNIT_Z);
+        s.camera.direction = Cartesian3.negate(Cartesian3.normalize(s.camera.position));
+
+        s.initializeFrame();
+        s.render();
+
+        waitsFor(function() {
+            render(s._context, s.frameState, s.globe);
+            return !equals(this.env, s._context.readPixels(), [0, 0, 0, 0]);
+        }, 'the central body to be rendered', 5000);
+    });
+
     it('renders with multipass OIT if MRT is available', function() {
         if (scene._context.drawBuffers) {
             var s = createScene();
@@ -422,7 +473,7 @@ defineSuite([
 
             s.initializeFrame();
             s.render();
-            var pixels = s.context.readPixels();
+            var pixels = s._context.readPixels();
             expect(pixels[0]).not.toEqual(0);
             expect(pixels[1]).toEqual(0);
             expect(pixels[2]).toEqual(0);